Remove plan & move function to utils module
